chaM3Leon-be/requirements.txt
django
gunicorn
psycopg2-binary
python-decouple
djangorestframework
drf-spectacular
django-cors-headers
openai
anthropic
google-genai
requests
mlflow
paramiko
scp
pyjwt
django-allauth
djangorestframework-simplejwt

